refactor: move data access queries to dedicated class
